General clean up of parameter usage.  No doubt this will break a few things.
git-svn-id: svn+ssh://svn.code.sf.net/p/dotproject/code/branches/stable_2@6149 c625438d-3aaf-4c84-a37d-fcf3134530f0

+++ dotproject/modules/admin/viewuser.php
@@ -14,7 +14,7 @@ if ($user_id != $AppUI->user_id
 $AppUI->savePlace();
 
 if (isset($_POST['show_form'])) {
-	$add_pwt = dPgetParam($_POST, 'add_pwt', 0);
+	$add_pwt = (int)dPgetParam($_POST, 'add_pwt', 0);
 	$AppUI->setState('addProjWithTasks', $add_pwt);
 } else {
 	$AppUI->setState('addProjWithTasks', false);
