+++ dotproject/modules/companies/do_company_aed.php
@@ -3,7 +3,7 @@ if (!defined('DP_BASE_DIR')) {
   die('You should not access this file directly.');
 }
 
-$del = dPgetParam($_POST, 'del', 0);
+$del = (int)dPgetParam($_POST, 'del', 0);
 $obj = new CCompany();
 $msg = '';
