+++ dotproject/modules/forums/post_message.php
@@ -6,7 +6,7 @@ if (!defined('DP_BASE_DIR')) {
 // Add / Edit forum
 $message_id = isset($_GET['message_id']) ? $_GET['message_id'] : 0;
 $message_parent = isset($_GET['message_parent']) ? $_GET['message_parent'] : -1;
-$forum_id = dPgetParam($_REQUEST, 'forum_id', 0);
+$forum_id = (int)dPgetParam($_REQUEST, 'forum_id', 0);
 
 $canEdit = getPermission('forums', 'edit', $forum_id);
