+++ dotproject/modules/links/do_link_aed.php
@@ -7,7 +7,7 @@ if (!defined('DP_BASE_DIR')) {
 $link_id = intval(dPgetParam($_POST, 'link_id', 0));
 $del = intval(dPgetParam($_POST, 'del', 0));
 
-$not = dPgetParam($_POST, 'notify', '0');
+$not = (int)dPgetParam($_POST, 'notify', '0');
 if ($not!='0') $not='1';
 
 $obj = new CLink();
