+++ dotproject/modules/system/addeditpref.php
@@ -6,7 +6,7 @@ if (!defined('DP_BASE_DIR')) {
 ##
 ## add or edit a user preferences
 ##
-$user_id = dPgetParam($_GET, 'user_id', '0');
+$user_id = (int)dPgetParam($_GET, 'user_id', '0');
 // Check permissions
 if (!$canEdit && $user_id != $AppUI->user_id) {
   $AppUI->redirect("m=public&a=access_denied");
